plane_Python/requirements.txt
google-generativeai 
PyPDF2
numpy
flask_cors